Add methods to handle month change and properties for the arrows color
When I was going to use the library I found very strange that it didn't have any method to change months programmatically. Then I was digging into the code to search for that and I found that this could be achieved changing the StartDate. So I am doing this PR to make this a little more easy and intuitive.
And as I was already doing this, I added two properties to change the color of the arrows.